+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Move from Scala 2.12 to Scala 2.13. (#5192)
 * Open Policy Agent authorizer updated to a new version supporting Scala 2.13. See the _Changes, deprecations and removals_ sections for more details. (#5192)
 * Allow a custom password to be set for SCRAM-SHA-512 users by referencing a secret in the `KafkaUser` resource
+* Add support for `tls-external` authentication to User Operator to allow management of ACLs and Quotas for TLS users with user certificates generated externally (#5249) 
 
 ### Changes, deprecations and removals